% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_isolates.R
\name{clean_isolates}
\alias{clean_isolates}
\title{Get just the largest fully-connected subgraph of a graph}
\usage{
clean_isolates(g)
}
\arguments{
\item{g}{An igraph object.}
}
\value{
An igraph object containing only the largest fully-connected subgraph of g
}
\description{
Get just the largest fully-connected subgraph of a graph
}
