% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleangraph.R
\name{cleangraph}
\alias{cleangraph}
\title{Clean a graph
JAN 6 2024 - DEPRECATED
Tidy up a graph that's been spit out of an automated process.}
\usage{
cleangraph(g)
}
\arguments{
\item{g}{An igraph object.}
}
\value{
An igraph object, constructed from the edgelist of g
}
\description{
Clean a graph
JAN 6 2024 - DEPRECATED
Tidy up a graph that's been spit out of an automated process.
}
