% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_graphs.R
\name{compare_graphs}
\alias{compare_graphs}
\title{cgraph - compare graphs}
\usage{
compare_graphs(g1.E, g2.E)
}
\arguments{
\item{g1.E}{An n * 2 character matrix: the edgelist for g1.}

\item{g2.E}{An m * 2 character matrix: the edgelist for g2.}
}
\value{
An integer vector of length n, one entry per row (edge) og g1.E.
The entry for each index is equal to the index of the row in g2.E where
the equivalent edge is found.
}
\description{
Check for similarities between graphs
}
\details{
Which of the edges of g1 can be found in g2, and to which edges do they correspond?
}
\examples{
data("simpleGs")
g1 <- as_edgelist(simpleGs$g1)
g2 <- as_edgelist(simpleGs$g2)

comp <- compare_graphs(g1,g2)

comp ## edge 1 of g1 corresponds to edge 1 of g2,
## edge 2 of g1 corresponds to edge 2 of g2, etc,
## edge 5 of g1 does not have a corresponding edge in g2

}
