% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_frequencies.R
\name{component_frequencies}
\alias{component_frequencies}
\title{component_frequencies}
\usage{
component_frequencies(charMat)
}
\arguments{
\item{charMat}{A character matrix created by the "morpho_matrix" function}
}
\value{
A list of two lists: "EdgeConsistency," the proportion of species in
which each edge is present, and VertexConsistency," the proportion of species
in which each vertex is present.
}
\description{
Vertex and Edge Consistency
}
\details{
How often does each edge and each vertex appear in a dataset?
}
\examples{
data("simpleGs")
mat <- morpho_matrix(simpleGs)
freq <- component_frequencies(mat)
freq

}
