% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/componentFrequencies.R
\name{componentFrequencies}
\alias{componentFrequencies}
\title{componentFrequencies}
\usage{
componentFrequencies(charMat)
}
\arguments{
\item{charMat}{A character matrix created by the "morphoMatrix" function}
}
\value{
A list of two lists: "EdgeConsistency," the proportion of species in
which each edge is present, and ScaleConsistency," the proportion of species
in which each vertex (scale) is present.
}
\description{
Vertex and Edge Consistency
}
\details{
How often does each edge and each vertex (scale) appear in a dataset?
}
\examples{
data("simpleGs")
mat <- morphoMatrix(simpleGs)
freq <- componentFrequencies(mat)
freq

}
