% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cycles.R
\name{find_cycles}
\alias{find_cycles}
\title{Find cycles in a graph
find and list all cycles (rings of vertices) in a graph larger than a given
number of vertices.}
\usage{
find_cycles(g, minlength = 5)
}
\arguments{
\item{g}{An igraph object.}

\item{minlength}{An integer; the length of the shortest desired cycle. Set by default to 5}
}
\value{
A list of cycles as vertex paths through the graph g
}
\description{
Find cycles in a graph
find and list all cycles (rings of vertices) in a graph larger than a given
number of vertices.
}
\examples{
find_cycles(simpleGs$g6)
find_cycles(simpleGs$g6, minlength = 4)
}
