% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_edge.R
\name{find_edge}
\alias{find_edge}
\title{Find an edge}
\usage{
find_edge(edge, edgelist)
}
\arguments{
\item{edge}{A character vector of length 2, the names of the vertices of the edge.}

\item{edgelist}{A n * 2 character matrix, an edgelist for a graph.}
}
\value{
a scalar, either the index of the row of edge in edgelist, or NA if
the edge cannot be found in the edgelist
}
\description{
Does a given edge exist in an edgelist?
}
\examples{
data("simpleGs")
find_edge(c('B','D'),as_edgelist(simpleGs$g1)) ##neither the edge "B-D" nor
## the edge "D-B" is present in g1
find_edge(c('B','D'),as_edgelist(simpleGs$g2) # the edge "B-D" (or the edge
## "D-B") is the 5th edge in g1

}
