% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sisteredges.R
\name{get_sisteredges}
\alias{get_sisteredges}
\title{get_sisteredges}
\usage{
get_sisteredges(input_matrix)
}
\arguments{
\item{input_matrix}{an n × 4 matrix where the rows correspond to
problematic edges in a graph and the 4 columns contain the 4 vertices which
complete the 2 triangles surrounding the problem edge}
}
\value{
A data frame of information about the groupings of problem edges
}
\description{
For a set of problematic edges and their vertices.
}
\details{
This function takes a graph and vertex names of an edge and returns
vertex names for the 2 vertices which complete the 2 adjacent triangles
}
\examples{
input <- matrix(
c("a","c","b","e",
"c","e","a","d"), nrow = 2, byrow = TRUE)
get_sisteredges(input)
}
