% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_edit_distance.R
\name{graph_edit_distance}
\alias{graph_edit_distance}
\title{graph_edit_distance}
\usage{
graph_edit_distance(g1, g2)
}
\arguments{
\item{g1}{An igraph object; the first graph.}

\item{g2}{An igraph object; the second graph.}
}
\value{
A list containing
\enumerate{
\item g1.df - A data frame of topological changes to g1 and how they were resolved
\item g2.df - A data frame of topological changes to g2 and how they were resolved
\item distances - a list containing
node - the number of un-shared vertices between g1 and g2
swap - the number of "edge swaps" between g1 and g2
edge - the number of non-"edge swap" unshared edges between g1 and g2
}
}
\description{
Graph edit distance, number of topological changes between two planar
networks with mostly homologous vertices. This does not measure differences
in vertex or edge attributes (i.e., edge weight differences).
}
\details{
This function takes two graphs and measures edit distance between them,
returning
}
\examples{
data("simpleGs")
dist <- graph_edit_distance(simpleGs$g1,simplegs$g2)
dist

}
