% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphEditDist.R
\name{graphEditDist}
\alias{graphEditDist}
\title{graphEditDist}
\usage{
graphEditDist(g1, g2)
}
\arguments{
\item{g1}{An igraph object; the first graph.}

\item{g2}{An igraph object; the second graph.}
}
\value{
A list containing
   1) g1.df - A data frame of topological changes to g1 and how they were resolved
   2) g2.df - A data frame of topological changes to g2 and how they were resolved
   3) distances - a list containing
       node - the number of un-shared vertices between g1 and g2
       swap - the number of "edge swaps" between g1 and g2
       edge - the number of non-"edge swap" unshared edges between g1 and g2
}
\description{
Graph edit distance, number of topological changes between two planar
networks with mostly homologous vertices. This does not measure differences
in vertex or edge attributes (i.e., edge weight differences).
}
\details{
This function takes two graphs and measures edit distance between them,
returning
}
\examples{
data("simpleGs")
dist <- graphEditDist(simpleGs$g1,simplegs$g2)
dist

}
