% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lizard_setup.R
\name{lizard_setup}
\alias{lizard_setup}
\title{Set up a graph with everything needed to plot out a comprehensible lizard.}
\usage{
lizard_setup(graph, firstscale = "rostral")
}
\arguments{
\item{graph}{An igraph object, the scale network to be set up.}

\item{firstscale}{A character scalar, the name of a vertex in graph that should
be used for anterior-posterior ordination. Usually, the most anterior scale
on the lizar's head. Therefore, this is set by default to 'rostral'}
}
\value{
An igraph object with the following vertex attributes:
   str - the strength of the vertex
   deg - the degree of the vertex
   scaletype - the general scale type the vertex corresponds to; the name, ignoring
       any ordination or enumeration
   pos - distance from the firstscale
   side - 0 for scales on the midline; 1 for Right side; -1 for left side
}
\description{
Set up a graph with everything needed to plot out a comprehensible lizard.
}
