% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphoMatrix.R
\name{morphoMatrix}
\alias{morphoMatrix}
\title{Generate a morphological character matrix from a list of graphs}
\usage{
morphoMatrix(gl)
}
\arguments{
\item{gl}{A list of graphs}
}
\value{
A n*e matrix, where n is the number of graphs in gl,
and e is the total number of unique edges in gl.
An element, \link{n1,e1} is the weight of edge e1 in graph n1.
}
\description{
Generate a morphological character matrix from a list of graphs
}
\examples{
data("simpleGs")
mat <- morphoMatrix(simpleGs)
mat

}
