% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborlist.R
\name{neighborlist}
\alias{neighborlist}
\title{List the neighboring vertices}
\usage{
neighborlist(v1, g)
}
\arguments{
\item{v1}{A character vector, a named vertex in a network}

\item{g}{An n * 2 character matrix: the edgelist for the network.}
}
\value{
A character vector naming all the vertices neighboring v1.
}
\description{
get a list of all vertices which v1 neighbors
}
