% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pholidosis_plot.R
\name{pholidosis_plot}
\alias{pholidosis_plot}
\title{pholidosis_plot}
\usage{
pholidosis_plot(g, title = NULL, body = TRUE, ...)
}
\arguments{
\item{g}{An igraph object: the pholidosis network you want to plot}

\item{title}{A character vector: the title for the graph}

\item{body}{A boolean. If true (default), this removes the irst two vertices
of the network, which should correspond to the "head" and "body" vertices.
This makesit easier to project the plot onto a flat surface.}
}
\value{
a ggplot plot of the pholidosis network
}
\description{
Provides a quick way to plot a readable pholidosis network.
}
\details{
This function uses the ggraph package to estimate a good layout for a
pholidosis network
different edges, and different edge weights.
}
\examples{
pth <- system.file("extdata", "DibamidaeDemo.xlsx", package = "pholidosis")
net <- excel_to_network(pth)
pholidosis_plot(net[[1]], title = "Anelytropsis papillosus pholidosis network")
}
