% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnet_distance.R
\name{pnet_distance}
\alias{pnet_distance}
\title{pnet_distance}
\usage{
pnet_distance(g1, g2, truncate = TRUE)
}
\arguments{
\item{g1}{An igraph object; the first graph.}

\item{g2}{An igraph object; the second graph.}

\item{truncate}{A logical scalar. If TRUE (default), the function returns only
the three distance measures. If false, it returns the distance measures and
the data frames built by graph_edit_distance.}
}
\value{
Either a named vector of length 3, containing:
\enumerate{
\item the weight difference of shared edges
\item the number of edge changes among shared scales
\item the number of unshared scales
between graphs g1 and g2.
OR
a list containing:
topo - the distance measures and data frames built by graph_edit_distance
measure - the named vector described above.
}
}
\description{
Measures the full graph edit distance between two graphs.
}
\details{
This function calls graph_edit_distance to get topology changes and also
compares edge weights, presenting the total number of different nodes,
different edges, and different edge weights.
}
\examples{
filepath <- system.file("extdata", "DibamidaeDemo.xlsx", package = "pholidosis")
Dib <- excel_to_network(filepath)
dist <- pnet_distance(Dib$Anelytropsis_papillosus,Dib$D_alfredi)
}
