% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_network.R
\name{scale_network}
\alias{scale_network}
\title{scale_network}
\usage{
scale_network(
  df,
  checkAsymmetry = FALSE,
  verbose = FALSE,
  setup_fun = lizard_setup
)
}
\arguments{
\item{df}{A data frame, the scale table.}

\item{checkAsymmetry}{A logical scalar, if TRUE, the function checks whether
df is a symmetric matrix}

\item{verbose}{A logical scalar. If TRUE, the function prints errors when they occur.}

\item{setup_fun}{A setup function th help plot the netowrk. See the "lizard_setup" function for an example}
}
\value{
An igraph object: If df is valid according to verify_matrix,
the pholidosis network. If df is not a valid matrix, an empty graph.
}
\description{
This loads a dataframe (df) of scale relationships as an igraph object
}
\details{
See excel_to_network for importing networks from excel sheets, lizard_setup
for handling the assignment of vertex attributes, and verify_matrix for
a system handling the individual problems with improperly formatted dataframes
or dataframes with bad (non-numeric) entries.

If you use the 1-2-3 unfused-partially fused-fused numeric scale in your
matrix, lizard_setup will change your edge weights to 1-1.5-2 to make scale
subtractions and scale fusions numerically equivalent for graph_exit_distance
measurement.
}
\examples{
library(readxl)
filepath <- system.file("extdata", "DibamidaeDemo.xlsx", package = "pholidosis")
anelytropsis.adj <- read_xlsx(filepath, sheet = 1, col_names = TRUE)
anelytropsis.net <- scale_network(anelytropsis.adj)

}
