% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importScaleNetwork.R
\name{scaleNetwork}
\alias{scaleNetwork}
\title{This loads a dataframe (df) of scale relationships as an igraph object}
\usage{
scaleNetwork(df, checkAsymmetry = FALSE, verbose = FALSE)
}
\arguments{
\item{df}{A data frame, the scale table.}

\item{checkAsymmetry}{A logical scalar, if TRUE, the function checks whether
df is a symmetric matrix}

\item{verbose}{A logical scalar. If TRUE, the function prints errors when they occur.}
}
\value{
An igraph object, the scale network.
}
\description{
This loads a dataframe (df) of scale relationships as an igraph object
}
