% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedneighbors.R
\name{sharedneighbors}
\alias{sharedneighbors}
\title{Do vertices share neighbors?}
\usage{
sharedneighbors(edge, g1.E)
}
\arguments{
\item{edge}{A character vector of length 2: an edge in a network.}

\item{g1.E}{An n * 2 character matrix: the edgelist for the network.}
}
\value{
A character vector of all the vertices that both vertices of edge connect to
}
\description{
Find  all the vertices that both vertices in an edge connect to.
}
