% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleGs.R
\docType{data}
\name{simpleGs}
\alias{simpleGs}
\title{Six simple scale networks}
\format{
A list of six graphs
}
\source{
code below
}
\usage{
data(simpleGs)
}
\description{
This dataset contains simple networks for use in demonstrating the "pholidosis" package
}
\examples{
data('simpleGs')

}
\keyword{datasets}
