% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/what_two_triangles.R
\name{what_two_triangles}
\alias{what_two_triangles}
\title{what_two_triangles}
\usage{
what_two_triangles(graph, v1, v2)
}
\arguments{
\item{v1}{A string: the name of the first vertex of the target edge}

\item{v2}{A string: the name of the second vertex of the target edge}

\item{g}{An igraph object; the graph in which the target edge exists}
}
\value{
A vector of two vertex names
}
\description{
In a fully triangulated graph, any edge is part of 2 triangles.
For problematic edges, we want to know which 2 triangles those are.
}
\details{
This function takes a graph and vertex names of an edge and returns
vertex names for the 2 vertices which complete the 2 adjacent triangles
}
\examples{
g <- graph_from_literal(A-B-C-D-A,A-C)
tri_tips<- what_two_triangles(g, "A","C")
tri_tips
}
