% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_to_network.R
\name{excel_to_network}
\alias{excel_to_network}
\title{excel_to_network}
\usage{
excel_to_network(filename, verbose = FALSE, ...)
}
\arguments{
\item{filename}{A character vector specifying the file path of the excel file.}

\item{verbose}{A logical vector. If TRUE, funciton prints progress and errors.}

\item{...}{other parameters, including a setup function for use by scaleNetwork()}
}
\value{
A list of igraph objects, one for each sheet in the file.
}
\description{
Import a scale network or many scale networks from an excel file.
}
\details{
Calls verifyMatrix to troubleshoot and creates igraph objects using
scaleNetwork.
}
\examples{

filepath <- system.file("extdata", "DibamidaeDemo.xlsx", package = "pholidosis")
Dib <- excel_to_network(filepath)
plot(Dib$Anelytropsis_papillosus) #check it out, it's Anelytropsis papillosus
}
