% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lizardSetup.R, R/lizard_setup.R
\name{lizard_setup}
\alias{lizard_setup}
\title{lizardSetup}
\usage{
lizard_setup(graph, firstscale = "rostral")

lizard_setup(graph, firstscale = "rostral")
}
\arguments{
\item{graph}{An igraph object, the scale network to be set up.}

\item{firstscale}{A character scalar, the name of a vertex in graph that should
be used for anterior-posterior ordination. Usually, the most anterior scale
on the lizar's head. Therefore, this is set by default to 'rostral'}
}
\value{
An igraph object with the following vertex attributes:
   str - the strength of the vertex
   deg - the degree of the vertex
   scaletype - the general scale type the vertex corresponds to; the name, ignoring
       any ordination or enumeration
   pos - distance from the firstscale
   side - 0 for scales on the midline; 1 for Right side; -1 for left side

An igraph object with the following vertex attributes:
   str - the strength of the vertex
   deg - the degree of the vertex
   scaletype - the general scale type the vertex corresponds to; the name, ignoring
       any ordination or enumeration
   pos - distance from the firstscale
   side - 0 for scales on the midline; 1 for Right side; -1 for left side
}
\description{
Set up a graph with everything needed to plot out a comprehensible lizard.
}
\details{
This assigns vertex properties based on scales and re-aligns edge properties
so that the edge weights c(1,2,3) become c(1,1.5,2)
}
\examples{
library(readxl)
filepath <- system.file("extdata", "DibamidaeDemo.xlsx", package = "pholidosis")
anelytropsis.adj <- read_excel(filepath, sheet = 1)

# verify that all column and row names have matches
anelytropsis.adj <- verifyMatrix(anelytropsis.adj) # they do,
# and now there are row names

anelytropsis.net <- scaleNetwork(anelytropsis.adj, lizard = FALSE)
vertex_attr(anelytropsis.net) #vertices only have a name attribute

anelytropsis.net <- lizardSetup(anelytropsis.net) #now each vertex has name
# strength, degree, scaletype, position, and side attributes!

library(readxl)
filepath <- system.file("extdata", "DibamidaeDemo.xlsx", package = "pholidosis")
anelytropsis.adj <- read_excel(filepath, sheet = 1)

# verify that all column and row names have matches
anelytropsis.adj <- verifyMatrix(anelytropsis.adj) # they do,
# and now there are row names

anelytropsis.net <- scaleNetwork(anelytropsis.adj, lizard = FALSE)
vertex_attr(anelytropsis.net) #vertices only have a name attribute

anelytropsis.net <- lizard_setup(anelytropsis.net) #now each vertex has name
# strength, degree, scaletype, position, and side attributes!

}
