% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnet_distance.R
\name{pnet_distance}
\alias{pnet_distance}
\title{Measures the full graph edit distance  between two graphs}
\usage{
pnet_distance(g1, g2, truncate = TRUE)
}
\arguments{
\item{g1}{An igraph object; the first graph.}

\item{g2}{An igraph object; the second graph.}

\item{truncate}{A logical scalar. If TRUE (default), the function returns only
the three distance measures. If false, it returns the distance measures and
the data frames built by graphEditDist.}
}
\value{
Either a named vector of length 3, containing:
  1) the weight difference of shared edges
  2) the number of edge changes among shared scales
  3) the number of unshared scales
  between graphs g1 and g2.
  OR
  a list containing:
  topo - the distance measures and data frames built by graphEditDist
  measure - the named vector described above.
}
\description{
Measures the full graph edit distance  between two graphs
}
