% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleNetwork.R
\name{scaleNetwork}
\alias{scaleNetwork}
\title{scaleNetwork}
\usage{
scaleNetwork(df, checkAsymmetry = FALSE, verbose = FALSE, lizard = TRUE)
}
\arguments{
\item{df}{A data frame, the scale table.}

\item{checkAsymmetry}{A logical scalar, if TRUE, the function checks whether
df is a symmetric matrix}

\item{verbose}{A logical scalar. If TRUE, the function prints errors when they occur.}

\item{lizard}{A logical scalar. If TRUE, vertex properties are assigned via
the "lizard_setup" function}
}
\value{
An igraph object, the scale network.
}
\description{
This loads a dataframe (df) of scale relationships as an igraph object
}
\details{
See excel_to_network for importing networks from excel sheets, lizard_setup
for handling the assignment of vertex attributes, and verifyMatrix for
a system handling the individual problems with improperly formatted dataframes
or dataframes with bad (non-numeric) entries
}
\examples{
library(readxl)
filepath <- system.file("extdata", "DibamidaeDemo.xlsx", package = "pholidosis")
anelytropsis.adj <- read_xlsx(filepath, sheet = 1, col_names = TRUE)
anelytropsis.net <- scaleNetwork(anelytropsis.adj)

}
