% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifyMatrix.R
\name{verifyMatrix}
\alias{verifyMatrix}
\title{Find problems in a scale table}
\usage{
verifyMatrix(df, verbose = TRUE)
}
\arguments{
\item{df}{A data frame, the scale table.}

\item{verbose}{A logical scalar. If TRUE, it prints progress & errors.}
}
\value{
A data frame, either the original (df), the original data frame modified
   to work with other pholidosis functions, or a dummy data frame "bad.input.matrix" =1
}
\description{
Find problems in a scale table
}
